requirements.txt
amqp==5.3.1
APScheduler==3.11.2
asgiref==3.11.1
billiard==4.2.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.11
django-allauth==65.14.3
django-jazzmin==3.0.3
django-jet-reboot==1.3.10
django-rest-framework==0.1.0
django-unfold==0.80.2
djangorestframework==3.16.1
kombu==5.6.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
psycopg2==2.9.11
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
